# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "Laravel API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8097/api",
            "description": "Local server"
        },
        {
            "url": "https://api.prod.example.com/api",
            "description": "Production server"
        },
        {
            "url": "https://api.staging.example.com/api",
            "description": "Staging server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "5d4fa5d5607ed675ba93907e1fc94534",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User has been registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login a user",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User has been logged successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the authenticated user's information",
                "operationId": "2baab1eeb5fd49516237a38da22e7a07",
                "responses": {
                    "200": {
                        "description": "Authenticated user info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh the access token",
                "operationId": "ba3d43e7c90e3d31ce13fca10f19b70f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshTokenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Refreshed token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "title": "Login Request",
                "description": "Request body for user login",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "The email address of the user",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "The password for the user",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
            "RefreshTokenRequest": {
                "title": "Refresh Token Request",
                "description": "Request body for refreshing an access token",
                "required": [
                    "refresh_token"
                ],
                "properties": {
                    "refresh_token": {
                        "description": "The refresh token issued to the user",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "Register Request",
                "description": "Request body for user registration",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "The name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email address of the user",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "The password for the user",
                        "type": "string",
                        "format": "password"
                    },
                    "password_confirmation": {
                        "description": "The password confirmation for the user",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
...
            "Token": {
                "title": "Token",
                "description": "Response body for access token",
                "required": [
                    "access_token",
                    "token_type",
                    "expires_in"
                ],
                "properties": {
                    "access_token": {
                        "description": "The access token issued to the user",
                        "type": "string"
                    },
                    "token_type": {
                        "description": "The type of the token issued",
                        "type": "string"
                    },
                    "expires_in": {
                        "description": "The duration in seconds for which the token is valid",
                        "type": "integer"
                    },
                    "refresh_token": {
                        "description": "The refresh token issued to the user",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the user",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email address of the user",
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "description": "The date and time when the user's email was verified",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "The date and time when the user was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the user was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://lureg.test/oauth/authorize",
                        "tokenUrl": "http://lureg.test/oauth/token",
                        "refreshUrl": "http://lureg.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}